server/requirements.txt
wheel
django
gunicorn
djangorestframework 
markdown  
django-filter
djangorestframework-simplejwt
jsonschema
django-cors-headers
psycopg2-binary
psycopg-binary
requests
redis
django-redis
sydney-py
